+++ lib/context_event_handler.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-class ContextEventHandler
-  # @interface method
-  def publish(context, event)
-    raise NotImplementedError.new("You must implement publish method.")
-  end
-end
+require_relative "context_publisher"
+
+# @deprecated Use ContextPublisher instead.
+ContextEventHandler = ContextPublisher
