+++ lib/default_context_publisher.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require_relative "context_publisher"
+
+class DefaultContextPublisher < ContextPublisher
+  attr_accessor :client
+
+  def initialize(client)
+    @client = client
+  end
+
+  def publish(context, event)
+    @client.publish(event)
+  end
+end
